% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_aggregation.R
\name{.aggregatePSM}
\alias{.aggregatePSM}
\title{Aggregate PSM to peptides}
\usage{
.aggregatePSM(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assay data to be
aggregated.}

\item{assays}{A integer vector specifying the indices of the assays to be
aggregated within the \code{QFeatures} object.}
}
\value{
A modified \code{QFeatures} object with new assays added.
The names of these new assays are prefixed with "peptides_".
}
\description{
Aggregate PSM to peptides
}
\keyword{internal}
