% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.divideReferenceTMT}
\alias{.divideReferenceTMT}
\title{Normalize TMT Data by Dividing by Reference Samples}
\usage{
.divideReferenceTMT(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be normalized.}

\item{assays}{A character vector specifying the names of the assays to be normalized.}
}
\value{
The \code{QFeatures} object with normalized assays based on reference samples.
}
\description{
Normalizes TMT data by dividing each sample by the corresponding reference
sample for normalization.
}
\keyword{internal}
