% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.filterAssayTMT}
\alias{.filterAssayTMT}
\title{Filter TMT Assays Based on Row Dimensions}
\usage{
.filterAssayTMT(qfeatures)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be filtered.}
}
\value{
The \code{QFeatures} object with only the assays that meet the row dimension criteria.
}
\description{
Filters TMT assays based on the number of features. Only assays with more than
120 features are retained.
}
\keyword{internal}
