% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.filterFDRTMT}
\alias{.filterFDRTMT}
\title{Filter TMT Data Based on False Discovery Rate (FDR)}
\usage{
.filterFDRTMT(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be filtered.}

\item{assays}{A character vector specifying the names of the assays to be filtered.}
}
\value{
The \code{QFeatures} object with filtered features based on FDR.
}
\description{
Filters TMT data based on q-values for proteins, using a 1\% FDR threshold.
}
\keyword{internal}
