% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.filterPSMTMT}
\alias{.filterPSMTMT}
\title{Filter TMT PSMs}
\usage{
.filterPSMTMT(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be filtered.}

\item{assays}{A character vector specifying the names of the assays to be filtered.}
}
\value{
The \code{QFeatures} object with filtered features for TMT data.
}
\description{
Filters PSMs (Peptide-Spectrum Matches) for TMT data based on specific criteria such as reverse sequences,
contaminants, and PIF (Precursor Ion Fraction) values.
}
\keyword{internal}
