% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.filterSCRTMT}
\alias{.filterSCRTMT}
\title{Filter TMT Samples Based on Sample Carrier Ratio (SCR)}
\usage{
.filterSCRTMT(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be filtered.}

\item{assays}{A character vector specifying the names of the assays to be filtered.}
}
\value{
The \code{QFeatures} object after filtering based on the SCR.
}
\description{
Computes and filters based on the Sample Carrier Ratio (SCR) for TMT data.
Samples are filtered out if their SCR exceeds 0.1.
}
\keyword{internal}
