% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{.generateDesignBaseTMT}
\alias{.generateDesignBaseTMT}
\title{Generate Experimental Design Data for 16-TMT}
\usage{
.generateDesignBaseTMT(nRuns)
}
\arguments{
\item{nRuns}{Integer. The number of runs in the experiment.}
}
\value{
A data.frame containing the base design data for 16-TMT.
}
\description{
This function generates base columns for design data used in a 16-TMT experiment.
}
\keyword{internal}
