% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{.generateDesignMetaTMT}
\alias{.generateDesignMetaTMT}
\title{Generate Additional Metadata for Experimental Design in 16-TMT}
\usage{
.generateDesignMetaTMT(nRuns)
}
\arguments{
\item{nRuns}{Integer. The number of runs in the experiment.}
}
\value{
A data.frame containing metadata for the design data in 16-TMT.
}
\description{
This function generates supplementary columns for the design data in a 16-TMT experiment.
}
\keyword{internal}
