% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{.generateDesignTMT}
\alias{.generateDesignTMT}
\title{Generate Experimental Design Data for 16-TMT}
\usage{
.generateDesignTMT(nRuns)
}
\arguments{
\item{nRuns}{Integer. The number of experimental runs.}
}
\value{
A data.frame containing the synthetic design data for 16-TMT.
}
\description{
This function creates synthetic experimental design data for a 16-TMT experiment.
}
\keyword{internal}
