% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{.generateQuantBaseTMT}
\alias{.generateQuantBaseTMT}
\title{Generate Base Quantitative Data for 16-TMT}
\usage{
.generateQuantBaseTMT(nRuns, nFeatures, naRate)
}
\arguments{
\item{nRuns}{Integer. The number of runs in the experiment.}

\item{nFeatures}{Integer. The number of features per run.}

\item{naRate}{Numeric. The rate of missing data (0-1) to apply to the quantitative data.}
}
\value{
A data.frame containing the base quantitative data for 16-TMT.
}
\description{
This function generates the core quantitative data, including reporter ion intensities
and feature-specific metadata, for a 16-TMT experiment.
}
\keyword{internal}
