% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{.generateQuantMetaTMT}
\alias{.generateQuantMetaTMT}
\title{Generate Additional Metadata for Quantitative Data in 16-TMT}
\usage{
.generateQuantMetaTMT(nRuns, nFeatures)
}
\arguments{
\item{nRuns}{Integer. The number of runs in the experiment.}

\item{nFeatures}{Integer. The number of features per run.}
}
\value{
A data.frame containing metadata for the quantitative data in 16-TMT.
}
\description{
This function generates supplementary columns for the quantitative data in a 16-TMT experiment.
}
\keyword{internal}
