% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_data_import.R
\name{.readQFeaturesLFDIA}
\alias{.readQFeaturesLFDIA}
\title{Read Quantitative LFDIA Data into QFeatures}
\usage{
.readQFeaturesLFDIA(quantitativeTable, designTable)
}
\arguments{
\item{quantitativeTable}{A data frame containing quantitative data.}

\item{designTable}{A data frame containing experimental design details,
which includes columns "runCol" and "quantCols".}
}
\value{
A \code{QFeatures} object containing \code{SingleCellExperiment} assays.
}
\description{
Utility function to read LFDIA data and convert it into a \code{QFeatures} object.
}
\keyword{internal}
