% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_data_import.R
\name{.readQFeaturesPlexDIA}
\alias{.readQFeaturesPlexDIA}
\title{Read Quantitative PlexDIA Data into QFeatures}
\usage{
.readQFeaturesPlexDIA(quantitativeTable, designTable, dataDIA)
}
\arguments{
\item{quantitativeTable}{A data frame containing quantitative data.}

\item{designTable}{A data frame containing experimental design details,
which includes columns "runCol" and "quantCols".}

\item{dataDIA}{A character string specifying the file path to the DIA data.}
}
\value{
A \code{QFeatures} object containing \code{SingleCellExperiment} assays.
}
\description{
Utility function to read PlexDIA data and convert it into a \code{QFeatures} object.
}
\keyword{internal}
