% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_data_import.R
\name{.readQFeaturesTMT}
\alias{.readQFeaturesTMT}
\title{Read Quantitative TMT Data into QFeatures}
\usage{
.readQFeaturesTMT(quantitativeTable, designTable, runCol)
}
\arguments{
\item{quantitativeTable}{A data frame containing quantitative data.}

\item{designTable}{A data frame containing experimental design details,
which includes columns "runCol" and "quantCols".}

\item{runCol}{A character string specifying the name of the column in the design table
that contains the runs or batches.}
}
\value{
A \code{QFeatures} object containing \code{SingleCellExperiment} assays.
}
\description{
Utility function to read TMT data and convert it into a \code{QFeatures} object.
}
\keyword{internal}
