% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_pre_processing.R
\name{.stepPreProcessingTMT}
\alias{.stepPreProcessingTMT}
\title{TMT Data Preprocessing Pipeline}
\usage{
.stepPreProcessingTMT(qfeatures, assays)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assays to be preprocessed.}

\item{assays}{A character vector specifying the names of the assays to be processed.}
}
\value{
The \code{QFeatures} object after preprocessing for TMT data.
}
\description{
Preprocesses TMT-type proteomic data by applying several filtering and normalization steps.
}
\keyword{internal}
