% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_management.R
\name{installScripts}
\alias{installScripts}
\title{Install All Scripts from benchmarkQFeatures}
\usage{
installScripts(destinationFolder = getwd())
}
\arguments{
\item{destination_folder}{A character string specifying the folder where the
scripts should be installed. Defaults to the current working directory.}
}
\description{
This function installs all R scripts located in the \code{inst/scripts} directory
of the benchmarkQFeatures package to a specified destination folder.
If no destination folder is provided, it defaults to the current working directory.
All installed scripts will be made executable.
}
\examples{
# Install all scripts to the current working directory
install_all_scripts()

}
