% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_benchmark.R
\name{leduc2022Benchmark}
\alias{leduc2022Benchmark}
\title{leduc2022 vignette benchmarking}
\usage{
leduc2022Benchmark(nCellRange)
}
\arguments{
\item{nCellRange}{range of number of cells to benchmark}
}
\value{
results
}
\description{
leduc2022 vignette benchmarking
}
