% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_benchmark.R
\name{leduc2022Generate}
\alias{leduc2022Generate}
\title{Generate leduc2022 data}
\usage{
leduc2022Generate(base, nCell)
}
\arguments{
\item{nCell}{number of cells to generate}
}
\value{
qfeatures
}
\description{
Generate leduc2022 data
}
