% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profilingWrapper}
\alias{profilingWrapper}
\title{Profile a Workflow Across Multiple Datasets}
\usage{
profilingWrapper(dataFolder, outputFolder, ...)
}
\arguments{
\item{dataFolder}{Character. Path to the folder containing subfolders with data files.
Each subfolder should contain \code{quant.csv} and \code{design.csv}.}

\item{outputFolder}{Character. Path where profiling output files will be saved.}

\item{...}{Additional arguments passed to the \code{profilingWorkflow} function.}
}
\value{
No return value. This function performs profiling and saves the output files.
}
\description{
This function profiles a workflow over multiple datasets by iterating through
folders containing quantitative and design CSV files. Profiling results are saved
for each dataset into separate output files.
}
\details{
This function iterates through each subfolder in \code{dataFolder}, checks for the presence of
\code{design.csv} and \code{quant.csv}, and then performs profiling of the specified workflow.
Profiling data is saved to files in \code{outputFolder}, named after each dataset folder.
}
\examples{
# To profile workflows across multiple datasets:
# profilingWrapper(dataFolder = "data", outputFolder = "profile_results")

}
