% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_benchmark.R
\name{renderBenchmarking}
\alias{renderBenchmarking}
\title{Title}
\usage{
renderBenchmarking(qfeaturesObject, nCell)
}
\arguments{
\item{qfeaturesObject}{qfeatures object}

\item{nCell}{number of cells}
}
\value{
time
}
\description{
Title
}
