% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_aggregation.R
\name{stepAggregate}
\alias{stepAggregate}
\title{Aggregate Features Over Assays}
\usage{
stepAggregate(qfeatures, assays, type)
}
\arguments{
\item{qfeatures}{A \code{QFeatures} object containing the assay data to be aggregated.}

\item{assays}{A integer vector specifying the indices of the assays to be
aggregated within the \code{QFeatures} object.}

\item{type}{A character string specifying the type of aggregation to be
performed c("PSM", "peptide").}
}
\value{
A modified \code{QFeatures} object with new assays added. The names of these
new assays are prefixed according to the type of aggregation used.
}
\description{
Aggregates features over specified assays in a \code{QFeatures} object.
It calculates the median of the values for each group across the
specified assays.
}
