% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_step_data_import.R
\name{stepDataImport}
\alias{stepDataImport}
\title{Data Import for Quantitative Proteomics}
\usage{
stepDataImport(quantitativePath, designPath, runCol, dataDIA = NULL, type)
}
\arguments{
\item{quantitativePath}{A character string specifying the file path to the table
containing quantitative data.}

\item{designPath}{A character string specifying the file path to the table with
experimental design details. The design table must contain columns "runCol"
and "quantCols" as required.}

\item{runCol}{A character string indicating the column in the design table that
specifies the runs or batches.}

\item{dataDIA}{A character string specifying the file path to DIA data, only required
if using "plexDIA" as the \code{type} parameter. Default is NULL.}

\item{type}{A character string specifying the type of quantitative data to import.
Accepted values are "TMT", "LFDIA", or "plexDIA".}
}
\value{
A \code{QFeatures} object containing \code{SingleCellExperiment} assays based on the specified
quantitative data type.
}
\description{
Imports quantitative proteomics data into a \code{QFeatures} object, compatible with
single-cell and multiplexed data types such as TMT, LFDIA, and plexDIA.
}
\examples{
# Importing TMT data
# qfeatures <- stepDataImport("quantData.csv", "designData.csv", "Run", type = "TMT")


}
