% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profiling.R
\name{profilingExecution}
\alias{profilingExecution}
\title{Execute Data Generation and Profiling Workflow}
\usage{
profilingExecution(
  nFeaturesRange = c(500, 1000),
  nRunsRange = c(4, 8),
  naRateRange = c(0.4, 0.6),
  nReplicates = 1,
  profilingTemp = "profilingTemp",
  outputFolder = "dataOutput/profilingResults"
)
}
\arguments{
\item{nFeaturesRange}{Integer vector. Specifies the range of the number of features to generate. Defaults to \code{c(100, 1000)}.}

\item{nRunsRange}{Integer vector. Specifies the range of the number of runs to generate. Defaults to \code{c(4, 8)}.}

\item{naRateRange}{Numeric vector. Specifies the range of the rate of missing values to introduce. Defaults to \code{c(0.4, 0.6)}.}

\item{nReplicates}{Integer. Number of replicate datasets to generate for each combination. Defaults to \code{1}.}

\item{profilingTemp}{Character. Path to a temporary folder for storing intermediate profiling data. Defaults to \code{"profilingTemp"}.}

\item{outputFolder}{Character. Path where profiling results will be saved. Defaults to \code{"dataOutput/profilingResults"}.}
}
\value{
No return value. This function performs data generation and profiling, saving results to \code{outputFolder}.
}
\description{
This function generates synthetic datasets for profiling based on specified parameters
and then performs profiling across these datasets. Temporary files generated during the process
are automatically cleaned up on exit.
}
\details{
The function first generates synthetic data for TMT profiling by calling \code{replicateData}.
Label-free DIA and plexed DIA generation functions can be added to extend functionality.
Next, \code{profilingWrapper} is called to execute profiling on the generated data.
Temporary files are cleaned up upon completion or if an error occurs.
}
\examples{
# Run profiling execution with default parameters:
# profilingExecution()

# Customize parameters:
# profilingExecution(nFeaturesRange = c(200, 500), nRunsRange = c(6, 12), naRateRange = c(0.3, 0.5))

}
